Add ordering within and between tasks. ELFF repo still busted.
In Puppet, ordering of statements and classes is NOT how resources are
provisioned. We have to explicitely order them with Stage and
'requires' and 'before'. Tedious, but "explicit is better than implicit."
y
Looks like Puppet invokes yum which fails on broken ELFF repo spec
(change DNS name) so we can NOT work around it by manipulating it with
our puppet provisioning. We can run it twice for now, or find/build a
custom Vbox with this fixed.